requirements.txt
fastapi
uvicorn
telethon
<<<<<<< HEAD
=======
python-dotenv
>>>>>>> ab3e4cc (Lokaler Build erfolgreich: FastAPI & Telethon laufen)
